[~] Машина состояний. Добавлена работа с сигналами и переход на другую страницу при создании документа в сервисном процессе
git-svn-id: http://svn.datateh.ru/alfrescoecm/Business%20Platform/trunk@2177 697f8a5b-bb4b-4f4c-ab90-ff1ad96ece97

+++ lecm-statemachine/repo/src/main/java/ru/it/lecm/statemachine/StateMachineHelper.java
@@ -667,7 +667,7 @@ public class StateMachineHelper implements StateMachineServiceBean {
         for (WorkflowDescriptor workflowDescriptor : workflowDescriptors) {
             if (currentExecutionId == null || !currentExecutionId.equals(workflowDescriptor.getExecutionId())) {
                 WorkflowInstance instance = serviceRegistry.getWorkflowService().getWorkflowById(workflowDescriptor.getExecutionId());
-                if (instance != null) {
+                if (instance != null && instance.isActive()) {
                     serviceRegistry.getWorkflowService().deleteWorkflow(workflowDescriptor.getExecutionId());
                 }
             }
